+++ test/test.ts
@@ -1,17 +1,16 @@
-import { Chyan } from "../lib";
-import { loader } from "../lib/core/loader";
-import path from "path";
-import { initializer } from "../lib/core/initializer";
-import { RouterStorage } from "../lib/core/router";
+// import { BootstrapApplication, ChyanApplication, Get } from "../lib";
+// import serve from "koa-static";
+// import { resolve } from "path";
 
-!(async function () {
-  const app = Chyan.createApplication();
-
-  await loader.load([path.resolve(__dirname, "controllers")]);
-
-  await initializer();
-
-  app.useRouter(RouterStorage.getRouter());
-
-  app.run(8033);
-})();
+// @ChyanApplication()
+// export class Application extends BootstrapApplication {
+//   main() {
+//     this.app.useRouter(this.router);
+//     this.app.useGlobalMiddleware(serve(resolve(process.cwd(), "public")));
+//     this.app.run(1234);
+//   }
+//   @Get()
+//   helloWorld() {
+//     return "Hello Chyan.co";
+//   }
+// }